+++ gym_art/quadrotor_multi/quadrotor_single.py
@@ -20,7 +20,6 @@ References:
 import argparse
 import logging
 import sys
-import time
 
 import gym_art.quadrotor_multi.get_state as get_state
 import gym_art.quadrotor_multi.quadrotor_randomization as quad_rand
