+++ gym_art/quadrotor_multi/scenarios/utils.py
@@ -37,6 +37,12 @@ QUADS_PARAMS_DICT = {
     'o_uniform_swarm_vs_swarm': [QUADS_FORMATION_LIST, [0.4, 0.8]],
 }
 
+def create_scenario(quads_mode, envs, num_agents, room_dims, room_dims_callback, rew_coeff, quads_formation,
+                    quads_formation_size):
+    cls = eval('Scenario_' + quads_mode)
+    scenario = cls(quads_mode, envs, num_agents, room_dims, room_dims_callback, rew_coeff, quads_formation,
+                   quads_formation_size)
+    return scenario
 
 def update_formation_and_max_agent_per_layer(mode):
     formation_index = np.random.randint(low=0, high=len(QUADS_PARAMS_DICT[mode][0]))
